Make window resizable from titlebar
Normally, a theme designer has two options: either let the window be resizeable from the titlebar at the cost of a thin, opaque stripe at the top of the window, or get rid of the stripe at the cost of not being able to resize the window from the titlebar anymore. This change will get rid of the stripe (i.e. cover it up) without sacrificing the ability to resize the window from the titlebar.